feat: 6, 7단계 미션 구현

+++ src/main/java/roomescape/Reservation.java
@@ -1,19 +1,19 @@
 package roomescape;
 
 public class Reservation {
-  private int id;
+  private Long id;
   private String name;
   private String date;
   private String time;
 
-  public Reservation(int id, String name, String date, String time) {
+  public Reservation(Long id, String name, String date, String time) {
     this.id = id;
     this.name = name;
     this.date = date;
     this.time = time;
   }
 
-  public int getId() {
+  public Long getId() {
     return this.id;
   }
   public String getName() {
@@ -26,7 +26,7 @@ public class Reservation {
     return this.time;
   }
 
-  public static Reservation toEntity(Reservation reservation, int id) {
+  public static Reservation toEntity(Reservation reservation, Long id) {
     return new Reservation(id, reservation.getName(), reservation.getDate(), reservation.getTime());
   }
 }
